requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain & RAG
langchain==0.1.6
langchain-google-genai==0.0.7
langchain-community==0.0.20

# Vector Store
faiss-cpu==1.7.4

# Google AI
google-generativeai==0.3.2

# Document Processing
beautifulsoup4==4.12.3
lxml==5.1.0
PyMuPDF==1.23.21
python-multipart==0.0.6

# Selenium
selenium==4.17.2
webdriver-manager==4.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0
